pyproject.toml
[project]
name = "model-rollback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.38.40",
 "joblib>=1.5.1",
 "python-dotenv>=1.1.0",
 "sagemaker>=2.92.0",
]

[tool.ruff]
line-length = 120
